ДЗ №3 на беспощадное ревью. Поработал над обработкой и выводом ошибок. Разделил логику и HTML.
